+++ packages/jest-reporters/src/get_result_header.ts
@@ -45,7 +45,7 @@ export default (
   const testDetail = [];
 
   if (result.perfStats?.slow) {
-    const runTime = (result.perfStats.end - result.perfStats.start) / 1000;
+    const runTime = result.perfStats.runtime / 1000;
 
     testDetail.push(LONG_TEST_COLOR(formatTime(runTime, 0)));
   }