+++ packages/jest-test-result/src/helpers.ts
@@ -58,6 +58,7 @@ export const buildFailureTestResult = (
   openHandles: [],
   perfStats: {
     end: 0,
+    runtime: 0,
     slow: false,
     start: 0,
   },
@@ -156,6 +157,7 @@ export const createEmptyTestResult = (): TestResult => ({
   openHandles: [],
   perfStats: {
     end: 0,
+    runtime: 0,
     slow: false,
     start: 0,
   },